fixed: images with lat/lon in EXIF are not displayed at their real position, but off by a meter or so
git-svn-id: http://josm.openstreetmap.de/svn/trunk@2909 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/gui/layer/geoimage/GeoImageLayer.java
@@ -501,7 +501,7 @@ public class GeoImageLayer extends Layer implements PropertyChangeListener {
 
         try {
             int deg;
-            float min, sec;
+            double min, sec;
             double lon, lat;
 
             Metadata metadata = JpegMetadataReader.readMetadata(e.file);
